Fix mixed precision training

+++ part_model/utils/loss.py
@@ -208,7 +208,7 @@ class SemiSumLoss(nn.Module):
             # Check whether target masks contain any negative number. We use
             # -1 to specify masks that we want to drop when seg_frac < 1.
             semi_mask = seg_targets[:, 0, 0] >= 0
-            seg_loss = torch.zeros_like(semi_mask, dtype=logits.dtype)
+            seg_loss = torch.zeros_like(semi_mask, dtype=torch.float32)
             seg_loss[semi_mask] = semi_seg_loss(seg_mask, seg_targets)
             loss += self.seg_const * seg_loss
         if self.reduction == "mean":
